backend/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
djoser==2.1.0
Pillow==10.4.0
PyJWT==2.1.0
django-filter==21.1
djangorestframework-simplejwt==4.7.2
drf-extra-fields==3.4.0